code for integration of messages
