+++ client/src/helpers/APICalls/createNewConversation.ts
@@ -1,18 +1,18 @@
 import { AuthApiData } from '../../interface/AuthApiData';
 import { FetchOptions } from '../../interface/FetchOptions';
 
-const createNewMessage = async (data: { text: string; conversationId: string }) => {
+const createNewConversation = async (data: { text: string; senderId: string; conversationId: string }) => {
   const fetchOptions: FetchOptions = {
     method: 'POST',
     headers: { 'Content-Type': 'application/json' },
     body: JSON.stringify(data),
     credentials: 'include',
   };
-  return await fetch(`/conversation/message`, fetchOptions)
+  return await fetch(`/conversation`, fetchOptions)
     .then((res) => res.json())
     .catch(() => ({
       error: { message: 'Unable to connect to server. Please try again' },
     }));
 };
 
-export default createNewMessage;
+export default createNewConversation;
