--- server/routes/conversation.js
@@ -1,19 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const protect = require("../middleware/auth");
-const {
-  createConversation,
-  sendMessage,
-  getAllConversations,
-  getAllMessages,
-} = require("../controllers/conversation");
-
-router.route("/create").post(protect,createConversation);
-
-router.route("/send").post(protect,sendMessage);
-
-router.route("/all").get(protect,getAllConversations);
-
-router.route("/messages/:conversationId").get(getAllMessages);
-
-module.exports = router;
